requirements.txt
# Core Dependencies
python-telegram-bot==20.7
aiohttp==3.9.1
python-dotenv==1.0.0

# Google AI/Gemini
google-generativeai==0.3.2

# HTTP and API
requests==2.31.0
httpx==0.25.2

# Data Processing
pandas==2.1.4
numpy==1.24.3

# Optional Enhancements (comment out for now if not used)
google-cloud-aiplatform==1.38.1
redis==5.0.1
apscheduler==3.10.4
pydantic==2.5.2
